requirements.txt
# ==== Core scientific stack ====
pandas==2.2.3
numpy==2.1.2

# ==== Statistics / ML ====
scipy
statsmodels
scikit-learn
linearmodels

# ==== Visualization & QA (optional but useful later) ====
matplotlib==3.9.2
plotly
seaborn
plotly
dash
fpdf2
tabulate

# ==== Country / geo normalization ====
pycountry==24.6.1

# ==== Time and date utils ====
python-dateutil==2.9.0.post0

# ==== Logging / YAML config (for extensibility) ====
PyYAML==6.0.2

# ==== (optional) Jupyter and notebooks ====
jupyterlab==4.2.5

# ==== (optional) Git integration helpers ====
GitPython==3.1.43
tqdm
joblib
tqdm
joblib

tqdm
joblib

# ==== Dev / CI / linting ====
pytest
black
isort

# ==== optional ====
img2pdf 
cairosvg
